+++ Package.swift
@@ -3,7 +3,7 @@ import PackageDescription
 let package = Package(
     name: "FluentPostgreSQL",
     dependencies: [
-        .Package(url: "https://github.com/qutheory/postgresql.git", majorVersion: 0, minor: 1),
-        .Package(url: "https://github.com/qutheory/fluent.git", majorVersion: 0, minor: 7)
+        .Package(url: "https://github.com/vapor/postgresql.git", majorVersion: 0, minor: 2),
+        .Package(url: "https://github.com/vapor/fluent.git", majorVersion: 0, minor: 10)
     ]
 )
