+++ Sources/FluentPostgreSQL/PostgreSQLSerializer.swift
@@ -3,12 +3,12 @@ import Fluent
 public final class PostgreSQLSerializer: GeneralSQLSerializer {
     var placeholderCount: Int = 0
 
-    public override func sql(_ value: Value) -> String {
+    public override func sql(_ value: Node) -> String {
         self.placeholderCount += 1
         return "$\(self.placeholderCount)"
     }
     
-    public override func serialize() -> (String, [Value]) {
+    public override func serialize() -> (String, [Node]) {
         self.placeholderCount = 0
         return super.serialize()
     }
@@ -16,35 +16,44 @@ public final class PostgreSQLSerializer: GeneralSQLSerializer {
     public override func sql(_ string: String) -> String {
         return "\(string)"
     }
-    
-    public override func sql(_ column: SQL.Column) -> String {
-        switch column {
-        case .primaryKey:
-            return sql("id") + " SERIAL PRIMARY KEY"
-        case .integer(let name):
-            return sql(name) + " INT"
-        case .string(let name, let length):
+
+    public override func sql(_ type: Schema.Field.DataType) -> String {
+        switch type {
+        case .id:
+            return "SERIAL PRIMARY KEY"
+        case .string(let length):
             if let length = length {
-                return sql(name) + " VARCHAR(\(length))"
+                return "VARCHAR(\(length))"
             } else {
-                return sql(name) + " VARCHAR(255)"
+                return "VARCHAR(255)"
             }
-        case .double(let name, _, _):
-            return sql(name) + " FLOAT"
+        case .double:
+            return "FLOAT"
+        default:
+            break
         }
+        return super.sql(type)
     }
     
-    public override func sql(_ data: [String: Value]) -> (String, [Value]) {
-        var clause: [String] = []
-        
-        var d = data
-        if d.keys.contains("id") {
-            d.removeValue(forKey: "id")
+    public override func sql(_ data: Node?) -> (String, [Node])? {
+        guard let node = data else {
+            return nil
         }
         
-        let values = Array(d.values)
+        guard case .object(var dict) = node else {
+            return nil
+        }
         
-        clause += sql(keys: Array(d.keys))
+        var clause: [String] = []
+        
+        if dict.keys.contains("id") {
+            
+            dict.removeValue(forKey: "id")
+        }
+        
+        let values = Array(dict.values)
+        
+        clause += sql(keys: Array(dict.keys))
         clause += "VALUES"
         clause += sql(values)
         
@@ -53,5 +62,6 @@ public final class PostgreSQLSerializer: GeneralSQLSerializer {
             values
         )
     }
-
 }
+
+
